requirements.txt
deep-translator==1.11.4
python-dotenv
psycopg2-binary
